+++ request.rest
@@ -2,8 +2,8 @@ POST http://localhost:3003/turma
 Content-Type: application/json
 
 {
-    "nome":"Romeu",
-    "modulo":"3"
+    "nome":"teste",
+    "modulo":"1"
 }
 
 ###
