+++ src/model/TurmaModel.ts
@@ -1,12 +1,12 @@
 export class TurmaModel{
-       constructor(
+      constructor(
       public id: string,
       public nome: string,
-      public modulo: "1"|"2"|"3"|"4"|"5"|"6"
+      public modulo: string 
     ){
        this.id = id,
        this.nome = nome,
        this.modulo = modulo
     }
  }
- 
+//  "1"|"2"|"3"|"4"|"5"|"6"? "ativo":"inativo"
